fixed styling for lists, listshow

+++ src/components/List.js
@@ -1,30 +1,42 @@
 import React from "react";
 import { Link } from "react-router-dom";
 import Button from "react-bootstrap/Button";
-import Card from "react-bootstrap/Card";
 import { ButtonGroup } from "react-bootstrap";
 
 export default function List(props) {
   function handleClick(event, list) {
+    const confirmed = window.confirm("Do you really want to delete this list?")
+    if (confirmed) {
     props.deleteList(list);
+    }
   }
 
   const { list } = props;
 
   return (
-    <Card className="list-card" id={list.id + "-list-card"}>
-      <h2>{list.name}</h2>
-      <ButtonGroup>
-      <Link className="btn btn-outline-primary" to={`/lists/${list.id}`}>
-        More Information
-      </Link>
-      <div className="">
-      <Button onClick={(e) => handleClick(e, list)} variant="outline-danger">
-        {" "}
-        Delete List
-      </Button>
-      </div>
-      </ButtonGroup>
-    </Card>
+    <div className="list-card">
+        <ButtonGroup size="sm" className="float-right p-2">
+          <Link
+            className="btn btn-custom list-btn btn-sm "
+            to={`/lists/${list.id}`}
+          >
+            More Information
+          </Link>
+          <Button
+            onClick={(e) => handleClick(e, list)}
+            variant="outline-danger"
+            className="list-btn btn-sm"
+          >
+            Delete
+          </Button>
+        </ButtonGroup>
+        <Link
+            className=" "
+            to={`/lists/${list.id}`}
+          >
+            <h3>{list.name}</h3>
+          </Link>
+      
+    </div>
   );
 }
